Add converstion of bananas to cups
It's tricky, different bananas are different sizes. Specifying cups
in theory ought to be more accurate, but I have little recent practice
with this, esp. since always switching to fresh (not frozen) bananas.

My refs are a 3 loaf version, which claims 4 cups (equivalent would be
2 2/3 for 2 loaves), and a 9 loaf version, which claims 10-15
cups (equivalent would be in the range 2.22 to 3.33). I was going to
say 2 3/4 based on the former, but decided to round up to 3.